+++ testsuite/testsuite_impl/testsuite.go
@@ -20,7 +20,7 @@ func NewTestsuite(serviceImage string) *Testsuite {
 
 func (suite Testsuite) GetTests() map[string]testsuite.Test {
 	return map[string]testsuite.Test{
-		"singleNodeNginxTest": DynamicSingleNodeExampleTest{ServiceImage: suite.serviceImage},
+		"singleNodeNginxTest": DynamicSingleNodeNginxTest{ServiceImage: suite.serviceImage},
 		"fixedSizeNginxTest":  FixedSizeNginxTest{ServiceImage: suite.serviceImage},
 	}
 }